config/requirements.txt
PyYAML==5.3.1
pandas==1.0.5
oauth2client
gspread==3.6.0
vk_api==11.7.0
docker==4.2.1
requests==2.24.0
google-api-python-client==1.9.3


